requirements.txt
upstash-vector==0.7.0
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
